Partial receive done

+++ database/migrations/2026_08_07_101732_create_product_partial_receive_histories_table.php
@@ -11,20 +11,27 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('product_partial_receive_histories', function (Blueprint $table) {
-            $table->id();
-            $table->foreign('transaction_id')->references('id')->on('transactions')->onDelete('cascade');
-            $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
-            $table->foreign('purchase_line_id')->references('id')->on('purchase_lines')->onDelete('cascade');
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');// user who received the product
-            $table->float('purchase_quantity')->comment('Ordered quantity from supplier');
-            //$table->float('already_received_quantity')->comment('Total quantity received from supplier');
-            $table->float('received_quantity')->comment('Partial quantity received from supplier');
-            $table->text('note')->nullable()->comment('Comment if any');
-            $table->dateTime('date')->comment('Date of product partial receive');
-            $table->timestamps();
-            $table->softDeletes();
-        });
+        if (!Schema::hasTable('product_partial_receive_histories')) {
+            Schema::create('product_partial_receive_histories', function (Blueprint $table) {
+                $table->increments('id');
+                $table->unsignedInteger('transaction_id');
+                $table->unsignedInteger('product_id');
+                $table->unsignedInteger('purchase_line_id');
+                $table->unsignedInteger('user_id');
+                $table->float('purchase_quantity')->comment('Ordered quantity from supplier');
+                $table->float('received_quantity')->comment('Partial quantity received from supplier');
+                $table->text('note')->nullable()->comment('Comment if any');
+                $table->dateTime('date')->comment('Date of product partial receive');
+                $table->timestamps();
+                $table->softDeletes();
+
+                // Add foreign keys
+                $table->foreign('transaction_id')->references('id')->on('transactions')->onDelete('cascade');
+                $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
+                $table->foreign('purchase_line_id')->references('id')->on('purchase_lines')->onDelete('cascade');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            });
+        }
     }
 
     /**
